新增rbac相关表，完成之前部分todo

+++ talktime-framework/talktime-service/src/main/java/com/qingmeng/config/netty/vo/WsOnlineOfflineNotifyVO.java
@@ -26,5 +26,5 @@ public class WsOnlineOfflineNotifyVO {
     /**
      * 在线人数
      */
-    private Long onlineNum;
+    private Integer onlineNum;
 }
